requirements.txt
langchain==0.0.309
python-dotenv==1.0.0
streamlit==1.27.2
streamlit_chat==0.1.1
openai==0.28.0
os
openai